Move command implementations into its own package.
